Refactor Fee, Member, and Payment controllers to improve namespace organization and add payment verification functionality; remove obsolete subscription view.
